+++ actix-http/CHANGES.md
@@ -1,6 +1,10 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
+### Added
+* HttpResponse builders for 1xx status codes. [#1768]
+
+[#1768]: https://github.com/actix/actix-web/pull/1768
 
 
 ## 2.1.0 - 2020-10-30